refactor: flatten stutter facade module names
Each subsystem facade module repeated its parent directory name
(e.g. Network.LibP2P.Multiaddr.Multiaddr, Switch.Switch, DHT.DHT).
Promote these 9 facades to the parent namespace so the module path no
longer stutters:

  Multiaddr.Multiaddr   -> Multiaddr
  Transport.Transport   -> Transport
  Switch.Switch         -> Switch
  DHT.DHT               -> DHT
  Identify.Identify     -> Identify
  Ping.Ping             -> Ping
  AutoNAT.AutoNAT       -> AutoNAT
  Relay.Relay           -> Relay
  DCUtR.DCUtR           -> DCUtR

Updates module headers, all import sites, and the cabal module lists.
Sibling submodules (Multiaddr.Codec, DHT.Lookup, etc.) are unchanged.
