+++ src/Network/LibP2P.hs
@@ -86,7 +86,7 @@ module Network.LibP2P
 import Network.LibP2P.Crypto.Ed25519 (generateKeyPair)
 import Network.LibP2P.Crypto.Key (KeyPair)
 import Network.LibP2P.Crypto.PeerId (PeerId, fromPublicKey, peerIdBytes, toBase58, parsePeerId, toCIDv1)
-import Network.LibP2P.Multiaddr.Multiaddr (Multiaddr (..), fromText, splitP2P, toText)
+import Network.LibP2P.Multiaddr (Multiaddr (..), fromText, splitP2P, toText)
 import Network.LibP2P.Multiaddr.Protocol (Protocol (..))
 import Network.LibP2P.MultistreamSelect.Negotiation (ProtocolId, StreamIO (..))
 import Network.LibP2P.Protocol.GossipSub.Handler
@@ -99,11 +99,11 @@ import Network.LibP2P.Protocol.GossipSub.Handler
   , stopGossipSub
   )
 import Network.LibP2P.Protocol.GossipSub.Types (GossipSubParams (..), defaultGossipSubParams)
-import Network.LibP2P.Protocol.Identify.Identify (registerIdentifyHandlers, requestIdentify)
-import Network.LibP2P.Protocol.Ping.Ping (PingError (..), PingResult (..), registerPingHandler, sendPing)
+import Network.LibP2P.Protocol.Identify (registerIdentifyHandlers, requestIdentify)
+import Network.LibP2P.Protocol.Ping (PingError (..), PingResult (..), registerPingHandler, sendPing)
 import Network.LibP2P.Switch.Dial (dial)
 import Network.LibP2P.Switch.Listen (ConnectionGater (..), defaultConnectionGater, switchListen, switchListenAddrs)
-import Network.LibP2P.Switch.Switch (addTransport, newSwitch, removeStreamHandler, setStreamHandler, switchClose)
+import Network.LibP2P.Switch (addTransport, newSwitch, removeStreamHandler, setStreamHandler, switchClose)
 import Network.LibP2P.Switch.Types (Connection (..), DialError (..), StreamHandler, Switch)
 import Network.LibP2P.Transport.TCP (newTCPTransport)
-import Network.LibP2P.Transport.Transport (Transport (..))
+import Network.LibP2P.Transport (Transport (..))
