+++ test/Test/Network/LibP2P/NAT/AutoNAT/AutoNATSpec.hs
@@ -8,11 +8,11 @@ import Control.Concurrent.STM (newTQueueIO, atomically, writeTQueue, readTQueue,
 import Data.Word (Word8)
 import Data.IORef (newIORef, readIORef, modifyIORef')
 import Network.LibP2P.NAT.AutoNAT.Message
-import Network.LibP2P.NAT.AutoNAT.AutoNAT
+import Network.LibP2P.NAT.AutoNAT
 import Network.LibP2P.MultistreamSelect.Negotiation (StreamIO (..))
-import Network.LibP2P.Multiaddr.Multiaddr (Multiaddr (..))
+import Network.LibP2P.Multiaddr (Multiaddr (..))
 import Network.LibP2P.Multiaddr.Protocol (Protocol (..))
-import Network.LibP2P.Multiaddr.Multiaddr (toBytes)
+import Network.LibP2P.Multiaddr (toBytes)
 import Network.LibP2P.Crypto.PeerId (PeerId (..))
 
 -- | Create an in-memory stream pair for testing.
